==> notebooks/2_join_data_M2.ipynb <==
## Is the high schools recruiting rating a good indicator of a high performaning athlete?  


# Uncomment to get the correct versions
from IPython.display import clear_output
!pip install -r ../requirements.txt
clear_output() 


import numpy as np
import pandas as pd
import altair as alt
import cfbd
import warnings 

pd.set_option('display.max_columns', None)
warnings.filterwarnings('ignore')


draft = pd.read_csv('../data/draft_2010_2024.csv')
recruits = pd.read_csv('../data/recruits_2008_2020.csv')
df_team_year = pd.read_csv('../data/team_season.csv') 

recruits.head() 

draft.head()

# Sometime there are two players with the same name - this will create duplicates
# About 2k of the 39k records are duplicated due to shared names.

merged_df = pd.merge(left = recruits, right = draft, how = 'left', left_on = 'name', right_on = 'Player')

col = ['name', 'rating', 'ranking', 'recruit_type', 'year', 'position'
     , 'height', 'weight','latitude', 'longitude', 'city','state_province','country', 'stars'
     , 'committed_to', 'athlete_id', 'Rnd', 'Pick', 'Player', 'draft_year', 'College/Univ']


merged_df = merged_df[col]
merged_df.head()

merged_df.shape

# Function to check if there are at least 4 sequential characters in common
# If the name of the school they committed to matches the name they were drafted from, then return a 1:
def has_common_sequence(str1, str2, min_seq_length=4):
    """
    This is a complicated function. 
    By joining from recruit data to draft data on the player's name, we create a many-to-many relationship.

    This occurs because some (but very few) recruits appear twice in the dataset (ex - Ron Smith), 
    and football players sometimes have the same name (ex - David Long). 

    This function first de-duplicates based on the combination of name and college of choice. 
    In other words we are assuming no two players at one school had the same name. 

    Next, we join to the draft dataset by looking for the player's name and college they attended, and the name and college they were drafted from. 

    More information can be found here in our milestone I project on slide 5:
    https://docs.google.com/presentation/d/1_CfHYqeOniPscvbb8VfQqQUgyf4xCSeC1spL_9M0ejw/edit#slide=id.g2b8248144f9_0_127

    """

    
    if pd.isna(str1) or pd.isna(str2):
        return 0

    for i in range(len(str1) - min_seq_length + 1):
        sequence = str1[i:i + min_seq_length]
        if sequence in str2:
            return 1
    return 0

# Apply the function to the DataFrame
merged_df['CommonSequence'] = merged_df.apply(lambda row: has_common_sequence(row['committed_to'], row['College/Univ']), axis=1)

merged_df['RN'] = merged_df.sort_values(['name', 'committed_to', 'CommonSequence'], ascending=[True, True,False]) \
                           .groupby(['name', 'rating']) \
                           .cumcount() + 1

#col = ['name', 'rating', 'stars', 'committed_to', 'athlete_id', 'Rnd', 'Pick', 'Player', 'draft_year', 'College/Univ', 'CommonSequence', 'RN']
#merged_df = merged_df[col]

# Examples of de-duplication logic in action

# merged_df[merged_df['name'] == 'Ron Smith']
# merged_df[merged_df['name'] == 'David Long']

# Remove duplicates

merged_df = merged_df[merged_df['RN'] == 1]
merged_df['is_drafted'] = np.where(merged_df['Rnd'].isna(), 0.0, 1.0)

# Include Supplement Draft 
# Players that were not elgibile to make it to the draft but they were drafted 
# reference document: https://en.wikipedia.org/wiki/List_of_NFL_supplemental_draft_picks#:~:text=The%20supplemental%20draft%20was%20enacted,enter%20the%20main%20NFL%20draft.
supplement_draft = ['Josh Brent', 'Harvey Unga', 'Josh Gordon','Terrelle Pryor', 'Isaiah Battle', 'Sam Beal', 'Adonis Alexander']

# Since there are two 'Jalen Thompson', we will have to manually set this to true
merged_df.loc[(merged_df['name'] == 'Jalen Thompson') & (merged_df['year'] == 2016), 'is_drafted'] = 1.0

# Mark these as drafted as well 
for player_sd in supplement_draft: 
    merged_df.loc[merged_df['name'] == player_sd, 'is_drafted'] = 1.0

# Get features about the school

df_teams = pd.read_csv('../data/teams.csv')
df_teams =  df_teams[['school', 'conference', 'latitude_school', 'longitude_school']]
df_teams.head()

# Add team features to our recruit-draft dataset
# NOTE: by making this an inner join, we are only including players who committed to a school, which we believe is the correct assumption. 

merged_df = pd.merge(left = merged_df, right = df_teams, left_on = 'committed_to', right_on = 'school', how = 'inner')

# Drop unneeded columns:
# drop_cols = ['name', 'recruit_type', 'athlete_id', 'Rnd', 'Pick', 'Player'
#              , 'draft_year', 'College/Univ', 'CommonSequence', 'RN', 'school']

drop_cols = ['recruit_type', 'Player', 'Rnd','College/Univ','school', 'Pick',  'draft_year',  'CommonSequence', 'RN']

merged_df = merged_df.drop(columns = drop_cols)

merged_df.sample(5)

    # eda_df = 
# merged_df[['athlete_id', 'Player']].isna().sum()
merged_df[['name','athlete_id', 'is_drafted', 'committed_to']].head() 
# merged_df[merged_df['name'] == 'Terrelle Pryor'].head() 
# merged_df[merged_df['school'] != merged_df['committed_to']].head() 
# eda_df[['school', 'committed_to']]

print('merged_df: ' + str(merged_df.shape))

# bring in dataset showing a team's 2-year window success
df_team_year['season'] = df_team_year['season'].fillna(0).astype(int) # one of the 2,000 records for the season column was null.. not sure why.
df_team_year = df_team_year[['school', 'season', 'wins_rolling_2year', 'games_played_rolling_2year', 'post_season_wins_rolling_2year', 'point_diff_rolling_2year', 'win_pct_rolling_2year']]
df_team_year[df_team_year['school'] == 'Michigan']

df_team_year[df_team_year['school'] == 'Washington State']

df_final = pd.merge(left = merged_df, right = df_team_year, how = 'left', left_on = ['committed_to', 'year'], right_on = ['school', 'season'])

# Game data begins in 2010.
# with a 2-year lag, the earliest year with data is 2012.
df_final = df_final[df_final['year'] >= 2012]
df_final.rename(columns={'city': 'hometown_city', 
                         'country': 'hometown_country'}, inplace=True)
df_final.head()

df_final.to_csv('../data/M2_Final.csv', index = False)
