==> notebooks/5_split_population.ipynb <==
## Split Test/Train Population

import pandas as pd 
from sklearn.model_selection import train_test_split


base_path = '../data/base_population.csv'

### Train/Test Split 
1. Run the clusters on the train dataset 
2. Predict the clusters 

df = pd.read_csv(base_path)
df.head() 

X = df.drop(['is_drafted'], axis=1)
y = df[['is_drafted']]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

print(X_train.shape[0])
print(X_test.shape[0])

base_path = '../data/train_test_split/'
X_train.to_csv(base_path+'x_train.csv')
X_test.to_csv(base_path+'x_test.csv')

y_train.to_csv(base_path+'y_train.csv')
y_test.to_csv(base_path+'y_test.csv')
